requirements.txt
panel
pandas
supabase
python-dotenv
requests
pillow

